added validation of the parameters of the request to get information about the city and the request to delete the city from the database

+++ src/main/java/site/alex/konon/sol/telegramBot/controller/CityController.java
@@ -1,6 +1,5 @@
 package site.alex.konon.sol.telegramBot.controller;
 
-import org.springframework.beans.factory.annotation.Autowired;
 import org.springframework.http.HttpStatus;
 import org.springframework.http.ResponseEntity;
 import org.springframework.web.bind.annotation.*;
@@ -9,6 +8,7 @@ import site.alex.konon.sol.telegramBot.repository.CityRepository;
 import org.slf4j.Logger;
 import org.slf4j.LoggerFactory;
 import site.alex.konon.sol.telegramBot.validator.CityValidator;
+import site.alex.konon.sol.telegramBot.validator.TextValidator;
 
 import javax.servlet.http.HttpServletRequest;
 import java.util.List;
@@ -43,6 +43,9 @@ public class CityController {
     @DeleteMapping("/city")
     public ResponseEntity deleteCity(@RequestParam(value = "city") String name,HttpServletRequest request) {
         logger.info("new del connect from ip {} , and city name is {}",request.getRemoteAddr(),name);
+        if (!TextValidator.noEmptyValidate(name)){
+            return new ResponseEntity("not valid",HttpStatus.BAD_REQUEST);
+        }
         City town = repository.findByName(name);
         if (town != null) {
             repository.delete(town);
@@ -71,6 +74,9 @@ public class CityController {
     @GetMapping("/city")
     public ResponseEntity findCity(@RequestParam(value = "city") String name, HttpServletRequest request) {
         logger.info("new get connect from ip {} , and city name is {}",request.getRemoteAddr(),name);
+        if(!TextValidator.noEmptyValidate(name)){
+            return new ResponseEntity("not valid",HttpStatus.BAD_REQUEST);
+        }
         City city = repository.findByName(name);
         if (city != null) {
             return new ResponseEntity(city.getText(), HttpStatus.OK);
@@ -80,7 +86,7 @@ public class CityController {
     }
 
     @GetMapping("/find")
-    public ResponseEntity find(@RequestParam(value = "city") String name, HttpServletRequest request) {
+    public ResponseEntity search(@RequestParam(value = "city") String name, HttpServletRequest request) {
         logger.info("new find connect from ip {} , and city name is {}",request.getRemoteAddr(),name);
         List<City> cities = repository.findByNameStartingWith(name);
         if (cities.size()>0){
